fix: AddOrUpdateAlarmController initializing ahead of time separate two controllers from inputTimeController acc to functionality
