+++ lib/app/modules/addOrUpdateAlarm/bindings/add_or_update_alarm_binding.dart
@@ -1,9 +1,8 @@
 import 'package:get/get.dart';
+import 'package:ultimate_alarm_clock/app/modules/addOrUpdateAlarm/controllers/alarm_input_time_controller.dart';
 import 'package:ultimate_alarm_clock/app/modules/bottomNavigationBar/controllers/bottom_navigation_bar_controller.dart';
 import 'package:ultimate_alarm_clock/app/modules/home/controllers/home_controller.dart';
 import 'package:ultimate_alarm_clock/app/modules/settings/controllers/settings_controller.dart';
-import 'package:ultimate_alarm_clock/app/modules/timer/controllers/timer_controller.dart';
-
 import '../controllers/add_or_update_alarm_controller.dart';
 
 class AddOrUpdateAlarmBinding extends Bindings {
@@ -18,8 +17,8 @@ class AddOrUpdateAlarmBinding extends Bindings {
     Get.lazyPut<SettingsController>(
       () => SettingsController(),
     );
-    Get.put<TimerController>(
-      TimerController(),
+    Get.lazyPut<AlarmInputTimeController>(
+      () => AlarmInputTimeController(),
     );
     Get.lazyPut<BottomNavigationBarController>(
       () => BottomNavigationBarController(),
